% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_values.R
\name{initial_values}
\alias{initial_values}
\title{initial values for Bayesian Elo steepness}
\usage{
initial_values(mat, rem = 0, chains = 4)
}
\arguments{
\item{mat}{square interaction matrix}

\item{rem}{numeric, the proportion of interactions to be removed if any.
Default is 0, i.e. all interactions are used.}

\item{chains}{the number of sets of initial values}
}
\value{
a list of length \code{chains}
}
\description{
initial values for Bayesian Elo steepness
}
\examples{
data("bonobos", package = "EloRating")
initial_values(bonobos)
}
