% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_bias.R
\name{interaction_bias}
\alias{interaction_bias}
\title{interaction bias in interaction matrix}
\usage{
interaction_bias(m, ranking_by = "DS", n_rand = 0, concise_results = TRUE)
}
\arguments{
\item{m}{input matrix (square)}

\item{ranking_by}{character, the ranking method to be used (so far, only
\code{"DS"} is supported)}

\item{n_rand}{numeric, number of randomizations (if \code{n_rand = 0}, the
the default, no randomizations are performed)}

\item{concise_results}{logical, if \code{TRUE} (default), no randomziation
estimates are returned. If \code{FALSE}, the results of the individual
randomizations are returned in the results}
}
\value{
a list
}
\description{
correlation between rank/score distance and interaction numbers and
coefficient of variation between interactions of individuals
}
\examples{
data(bonobos)
interaction_bias(bonobos, "DS", 100)

res <- interaction_bias(bonobos, "DS", 1000, FALSE)
hist(res$randos[, "ordinal_cor"])
abline(v = res$output$value[1], col = "red")
}
