% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bias.R
\name{plot_bias}
\alias{plot_bias}
\title{illustrate rank-distance bias in interaction matrix}
\usage{
plot_bias(x)
}
\arguments{
\item{x}{results of \code{\link{simple_steep_gen}}}
}
\value{
a plot
}
\description{
illustrate rank-distance bias in interaction matrix
}
\examples{
x <- simple_steep_gen(n_ind = 20, n_int = 1000, steep = 0.8, id_bias = 0, rank_bias = 1)
plot_bias(x)
interaction_bias(x$matrix)
x <- simple_steep_gen(n_ind = 20, n_int = 5000, steep = 0.8, id_bias = 0, rank_bias = 0)
plot_bias(x)
interaction_bias(x$matrix)
}
