% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dyads.R
\name{remove_dyads}
\alias{remove_dyads}
\title{remove interactions from matrix to increase sparseness}
\usage{
remove_dyads(m, by_interaction = FALSE, stop_at = 0.5, max_out = NULL)
}
\arguments{
\item{m}{input matrix}

\item{by_interaction}{logical, should interactions be removed interaction by
interaction, or by removing one dyad entirely at a time}

\item{stop_at}{numeric, fractions of unknown relationships to be reached}

\item{max_out}{numeric, the number of matrices to be returned maximally.
This is useful if the input matrix is fairly large. If set, this
will return the input matrix plus \code{max_out} randomly selected
matrices from the remaining produced matrices. So in fact, the output
comprises \code{max_out + 1} matrices (subject to the \code{stop_at}
specification).}
}
\value{
a list
}
\description{
remove interactions from matrix to increase sparseness
}
\examples{
data(bonobos)
res <- remove_dyads(bonobos)
length(res$matrices)

res <- remove_dyads(bonobos, max_out = 2)
# first plus two randomly selected = 3 matrices
length(res$matrices)
res$summary
}
