% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_steep_gen_alternating.R
\name{simple_steep_gen_alternating}
\alias{simple_steep_gen_alternating}
\title{generate dominance matrix with specified steepness}
\usage{
simple_steep_gen_alternating(n_ind, n_int, steeps, id_bias = TRUE)
}
\arguments{
\item{n_ind}{integer, the number of individuals}

\item{n_int}{integer, the number of interactions}

\item{steeps}{numeric (between 0 and 1), the desired steepness values. The
length of this vector determines how many periods will be generated.}

\item{id_bias}{logical, should some individuals be more observed than others
(default is \code{TRUE}). Can also be a numeric of the same length as
\code{n_ind}.}
}
\value{
a list with the first item being the interactions in sequence form
and the second item as matrix
}
\description{
generate dominance matrix with specified steepness
}
\examples{
simple_steep_gen_alternating(n_ind = 3, n_int = 10, steeps = c(0.3, 0.9))
}
