% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steep_anidom.R
\name{steep_anidom}
\alias{steep_anidom}
\title{steepness via repeatability (cf aniDom package)}
\usage{
steep_anidom(xdata, reps = 1000, normprob = FALSE)
}
\arguments{
\item{xdata}{matrix or eloobject}

\item{reps}{number of randomized sequences}

\item{normprob}{logical, use exponential curve (default) or normal curve
for calculating winning probabilities}
}
\value{
a steepness value
}
\description{
steepness via repeatability (cf aniDom package)
}
\examples{
data(bonobos, package = "EloRating")
steep_anidom(bonobos, reps = 20)
}
\references{
Sanchez-Tojar et al 2018
}
