% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steepness_elo.R
\name{steepness_elo}
\alias{steepness_elo}
\title{steepness based on winning probabilities from Elo-rating}
\usage{
steepness_elo(
  x,
  date = NULL,
  sliding = 0,
  n_rand = 1000,
  return_mean = TRUE,
  fixed_ranks = FALSE
)
}
\arguments{
\item{x}{either a matrix or the results of \code{\link{elo.seq}}}

\item{date}{\code{Date} or character (YYYY-MM-DD), optional date to extract
steepness on when \code{x} is the result of \code{\link{elo.seq}}.
Default is \code{NULL}, i.e. the last date in the sequence. Ignored
if \code{x} is a matrix.}

\item{sliding}{numeric, optional number of days centered around \code{date}
over which steepness is averaged. Default is \code{0}, only the
ratings on the specified date are used.}

\item{n_rand}{numeric, the number of randomized sequences generated if
\code{x} is a matrix. Otherwise ignored. Default is \code{1000}.}

\item{return_mean}{logical, should the mean of the randomizations be
returned or the individual values from randomizations or sliding}

\item{fixed_ranks}{logical, should the same ranking be used for the
regression (as derived by David's score)}
}
\value{
an (average) steepness value ranging between 0 and 1.
}
\description{
steepness based on winning probabilities from Elo-rating
}
\examples{
library(EloRating)
data(bonobos)
# from a matrix with ranomized sequences
steepness_elo(bonobos)
steepness_elo(bonobos, fixed_ranks = TRUE)

# from sequence
data(adv)
e <- elo.seq(adv$winner, adv$loser, adv$Date)
steepness_elo(e)
steepness_elo(e, date = "2010-01-12")
steepness_elo(e, date = "2010-01-12", sliding = 3)
}
