% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steepness_h.R
\name{steepness_h}
\alias{steepness_h}
\title{H steepness}
\usage{
steepness_h(mat, rand = 1000)
}
\arguments{
\item{mat}{matrix}

\item{rand}{number of randomized matrices (default is 1000)}
}
\value{
a steepness value
}
\description{
H steepness
}
\examples{
data(bonobos)
steepness_h(bonobos)
}
