% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob_alt.R
\name{winprob_alt}
\alias{winprob_alt}
\title{winning probabilities with individual performance spread}
\usage{
winprob_alt(r1, r2, s1 = 200, s2 = 200, do_plot = TRUE)
}
\arguments{
\item{r1, r2}{numeric, the ratings of two individuals}

\item{s1, s2}{numeric, the standard deviation of the individual performance spread. Must be positive. Its default, 200, corresponds to the standard.}

\item{do_plot}{logical, if \code{TRUE} produces a density plot of the two performance ratings and a plot of the distributions of the rating difference}
}
\value{
a winning probability for the first individual and optionally a plot
}
\description{
winning probabilities with individual performance spread
}
\examples{
winprob_alt(1200, 1000)
winprob(1200, 1000)

winprob_alt(1200, 1400)
winprob(1200, 1400)

winprob_alt(1200, 1400, 100, 250)
winprob_alt(1200, 1400, 50, 250)
}
