% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_ids.R
\name{zero_ids}
\alias{zero_ids}
\title{individuals without interactions}
\usage{
zero_ids(mat, return_ids = FALSE)
}
\arguments{
\item{mat}{a square matrix}

\item{return_ids}{logical, if \code{FALSE} (default) the number of
individuals is returned, otherwise the row/column indices or
column names if present}
}
\value{
numeric
}
\description{
individuals without interactions
}
\examples{
zero_ids(bonobos)
zero_ids(bonobos, return_ids = TRUE)
m <- matrix(nrow = 4, ncol = 4, 0)
m[1, 3] <- 1
zero_ids(m)
zero_ids(m, return_ids = TRUE)
}
