% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_steep_gen_alternating.R
\name{simple_steep_gen_alternating}
\alias{simple_steep_gen_alternating}
\title{generate dominance matrix with specified steepness}
\usage{
simple_steep_gen_alternating(n_ind, n_int, steeps, id_bias = 0, rank_bias = 0)
}
\arguments{
\item{n_ind}{integer, the number of individuals}

\item{n_int}{integer, the number of interactions}

\item{steeps}{numeric (between 0 and 1), the desired steepness values. The
length of this vector determines how many periods will be generated.}

\item{id_bias}{numeric, between 0 and 1. If 0 all individual are equally
likely to interact. If 1, some individuals have higher propensities
to interact}

\item{rank_bias}{numeric, between 0 and 1. If 0 there is no relationship
between rank distance and interaction propensity. If 1 there is a
strong relationship: dyads closer in rank interact more often.}
}
\value{
a list with the first item being the interactions in sequence form
and the second item as matrix
}
\description{
generate dominance matrix with specified steepness
}
\details{
In essence, this is just a wrapper for
\code{\link{simple_steep_gen}}.
}
\examples{
simple_steep_gen_alternating(n_ind = 3, n_int = 10, steeps = c(0.3, 0.9))
}
